+++ crispr_read_counts/utils.py
@@ -2,6 +2,7 @@ import gzip
 import re
 from contextlib import contextmanager
 import os
+import sys
 
 PLASMID_COUNT_HEADER = re.compile(r'^sgRNA\tgene', flags=re.I)
 DNA_PATTERN = re.compile(r'^[ATGC]+$', flags=re.I)
@@ -43,25 +44,24 @@ def open_plain_or_gzipped_file(file: str):
         f.close()
 
 
-def check_file_readable(fn):
-  if os.path.isfile(fn):
-    return os.access(fn, os.R_OK)
-  else:
-    return False
+def check_file_readable(fn, msg_if_fail=None):
+  result: bool = os.path.isfile(fn) and os.access(fn, os.R_OK)
 
-
-def check_file_writable(fn):
-  '''the function is clearly borrowed from web'''
-  if os.path.exists(fn):
-    # path exists
-    if os.path.isfile(fn):  # is it a file or a dir?
-      # also works when file is a link and the target is writable
-      return os.access(fn, os.W_OK)
+  if not result:
+    if msg_if_fail:
+      sys.exit(error_msg(msg_if_fail))
     else:
-      return False  # path is a dir, so cannot write as a file
-  # target does not exist, check perms on parent dir
-  pdir = os.path.dirname(fn)
-  if not pdir:
-    pdir = '.'
-  # target is creatable if parent dir is writable
-  return os.access(pdir, os.W_OK)
+      sys.exit()
+
+
+def check_file_writable(fn, msg_if_fail=None):
+  result: bool = (
+    (os.path.isfile(fn) and os.access(fn, os.W_OK)) if os.path.exists(fn) else
+    os.access(os.path.dirname(fn) or '.', os.W_OK)
+  )
+
+  if not result:
+    if msg_if_fail:
+      sys.exit(error_msg(msg_if_fail))
+    else:
+      sys.exit()